[수정] 학과별공지 코드를 전용 Repository로 옮김
