+++ app/src/main/java/com/ku_stacks/ku_ring/repository/NoticeRepository.kt
@@ -16,7 +16,6 @@ interface NoticeRepository {
     suspend fun updateSavedStatus(articleId: String, category: String, isSaved: Boolean)
     suspend fun updateNoticeToBeReadOnStorage(articleId: String, category: String)
     suspend fun clearSavedNotices()
-    fun getDepartmentNotices(shortName: String): Flow<PagingData<Notice>>
     fun deleteAllNoticeRecord()
     fun deleteSharedPreference()
 }
